requirements.txt
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
cffi==1.14.1
click==7.1.2
cryptography==2.9.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fuzzywuzzy==0.18.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.0
packaging==20.4
passlib==1.7.2
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-Levenshtein==0.12.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.1